+++ test/integration/cat/test_cat.py
@@ -8,18 +8,18 @@ import pytest
 from utils import run_command_and_get_output
 
 
-@pytest.mark.parametrize("args, found", [
+@pytest.mark.parametrize("args, expected_error", [
     (
         # Testing with unset .xlog or .snap file.
-        (),
+        "",
         "it is required to specify at least one .xlog or .snap file",
     ),
     (
         "path-to-non-existent-file",
-        "No such file or directory",
+        "error: could not collect WAL files",
     ),
 ])
-def test_cat_args_tests_failed(tt_cmd, tmp_path, args, found):
+def test_cat_args_tests_failed(tt_cmd, tmp_path, args, expected_error):
     # Copy the .xlog file to the "run" directory.
     test_xlog_file = os.path.join(os.path.dirname(__file__), "test_file", "test.xlog")
     test_snap_file = os.path.join(os.path.dirname(__file__), "test_file", "test.snap")
@@ -30,10 +30,10 @@ def test_cat_args_tests_failed(tt_cmd, tmp_path, args, found):
     cmd.extend(args)
     rc, output = run_command_and_get_output(cmd, cwd=tmp_path)
     assert rc == 1
-    assert found in output
+    assert expected_error in output
 
 
-@pytest.mark.parametrize("args, found", [
+@pytest.mark.parametrize("args, expected", [
     (
         ("test.snap", "--show-system", "--space=320", "--space=296", "--from=423", "--to=513"),
         ("lsn: 423", "lsn: 512", "space_id: 320", "space_id: 296"),
@@ -48,7 +48,7 @@ def test_cat_args_tests_failed(tt_cmd, tmp_path, args, found):
          'Result of cat: the file "test.snap" is processed below'),
     ),
 ])
-def test_cat_args_tests_successed(tt_cmd, tmp_path, args, found):
+def test_cat_args_tests_successed(tt_cmd, tmp_path, args, expected):
     # Copy the .xlog file to the "run" directory.
     test_xlog_file = os.path.join(os.path.dirname(__file__), "test_file", "test.xlog")
     test_snap_file = os.path.join(os.path.dirname(__file__), "test_file", "test.snap")
@@ -59,7 +59,30 @@ def test_cat_args_tests_successed(tt_cmd, tmp_path, args, found):
     cmd.extend(args)
     rc, output = run_command_and_get_output(cmd, cwd=tmp_path)
     assert rc == 0
-    for item in found:
+    for item in expected:
+        assert item in output
+
+
+@pytest.mark.parametrize("args, expected", [
+    (
+        ("test_file/test.xlog", "test_file/test.snap", "test_file"),
+        ('Result of cat: the file "test_file/test.xlog" is processed below',
+         'Result of cat: the file "test_file/test.snap" is processed below',
+         'Result of cat: the file "test_file/timestamp.snap" is processed below',
+         'Result of cat: the file "test_file/timestamp.xlog" is processed below'),
+    ),
+])
+def test_cat_directories_tests_successed(tt_cmd, tmp_path, args, expected):
+    # Copy files to the "run" directory.
+    test_src = os.path.join(os.path.dirname(__file__), "test_file")
+    test_dir = os.path.join(tmp_path, "test_file")
+    shutil.copytree(test_src, test_dir)
+
+    cmd = [tt_cmd, "cat"]
+    cmd.extend(args)
+    rc, output = run_command_and_get_output(cmd, cwd=tmp_path)
+    assert rc == 0
+    for item in expected:
         assert item in output